+++ tests/test_functions.cpp
@@ -343,6 +343,15 @@ NB_MODULE(test_functions_ext, m) {
 
     m.def("test_set_contains", [](nb::set s, nb::handle h) { return s.contains(h); });
 
+    m.def("test_frozenset", []() {
+        return nb::frozenset(nb::make_tuple("123", 123));
+    });
+
+    m.def("test_frozenset_contains", [](nb::frozenset s, nb::handle h) {
+        return s.contains(h);
+    });
+
+
     m.def("test_del_list", [](nb::list l) { nb::del(l[2]); });
     m.def("test_del_dict", [](nb::dict l) { nb::del(l["a"]); });